refactor as own middleware
