+++ src/handlers/http/middleware.rs
@@ -28,7 +28,6 @@ use actix_web::{
 use futures_util::future::LocalBoxFuture;
 
 use crate::{
-    audit::AuditLogBuilder,
     handlers::{
         AUTHORIZATION_KEY, KINESIS_COMMON_ATTRIBUTES_KEY, LOG_SOURCE_KEY, LOG_SOURCE_KINESIS,
         STREAM_NAME_HEADER_KEY,
@@ -46,16 +45,16 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
-struct Message {
-    common_attributes: CommonAttributes,
+pub struct Message {
+    pub common_attributes: CommonAttributes,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-struct CommonAttributes {
+pub struct CommonAttributes {
     #[serde(rename = "Authorization")]
     authorization: String,
     #[serde(rename = "X-P-Stream")]
-    x_p_stream: String,
+    pub x_p_stream: String,
 }
 
 pub trait RouteExt {
@@ -138,7 +137,6 @@ where
         For requests made from other clients, no change.
 
          ## Section start */
-        let mut stream_name = None;
         if let Some(kinesis_common_attributes) =
             req.request().headers().get(KINESIS_COMMON_ATTRIBUTES_KEY)
         {
@@ -156,27 +154,14 @@ where
                 HeaderName::from_static(LOG_SOURCE_KEY),
                 header::HeaderValue::from_static(LOG_SOURCE_KINESIS),
             );
-            stream_name.replace(message.common_attributes.x_p_stream);
-        } else if let Some(stream) = req.match_info().get("logstream") {
-            stream_name.replace(stream.to_owned());
-        } else if let Some(value) = req.headers().get(STREAM_NAME_HEADER_KEY) {
-            if let Ok(stream) = value.to_str() {
-                stream_name.replace(stream.to_owned());
-            }
         }
 
         /* ## Section end */
 
         let auth_result: Result<_, Error> = (self.auth_method)(&mut req, self.action);
 
-        // Ensures that log will be pushed to subscriber on drop
-        let mut log_builder = AuditLogBuilder::default();
-        log_builder.set_stream_name(stream_name.unwrap_or_default());
-        log_builder.update_from_http(&mut req);
         let fut = self.service.call(req);
         Box::pin(async move {
-            log_builder.set_deployment_id().await;
-
             match auth_result? {
                 rbac::Response::UnAuthorized => return Err(
                     ErrorForbidden("You don't have permission to access this resource. Please contact your administrator for assistance.")
@@ -186,22 +171,8 @@ where
                 ),
                 _ => {}
             }
-            let res = fut.await;
 
-            // Capture status_code and error information from response
-            match &res {
-                Ok(res) => {
-                    let status = res.status();
-                    log_builder.response.status_code = status.as_u16();
-                    // Use error information from reponse object if an error
-                    if let Some(err) = res.response().error() {
-                        log_builder.set_response_error(err.to_string());
-                    }
-                }
-                Err(err) => log_builder.set_response_error(err.to_string()),
-            }
-
-            res
+            fut.await
         })
     }
 }
